[#5] Fix usage of settings in efaLive daemon test
The test for the efaLive dameon used the real efaLive settings
implementation. This is an issue, since the settings check a path any
try to create it, if it does not exist. Since the test should test the
daemon, the settings can savely be mocked.

Added some file types to the ignore file.