[#5] Fix mocks in efaLive daemon tests
The assertions for calls of methods from the mocks were wrong. The
assertions basically never failed. You can call any method for a mock
and it will never report an error. In this case i used
'assert_called_once', which is not an existing method of a mock. It came
from some site via copy and paste. Since 'assert_called_once' is not in
the mock implementation, it is assumed that this method should be
mocked. So this call just passes without any issue.
The best solution from my point of view is to write your own assertions
and use call_count and call_args.