[#11] Implement work-around to set hardware clock
Implement a simple work-around to set the hardware clock to the system
time during shutdown. Normally this should be done by Debian, but there
is a bug in Systemd that prevents it. The work-around works for a normal
shutdown via efa close only. All other ways to shutdown the system are
not handled.